git-svn-id: https://svn.code.sf.net/p/highrisedev/code@80 f146f8d7-247d-460d-a1b3-d0a612191574

+++ Tower/floorBase.h
@@ -45,9 +45,9 @@ public:
    inline float GetY() { return mY; }
    inline float GetZ() {return mZ; }
 
-   inline void SetX(int x) { mX = x; }
-   inline void SetX2(int x) { mX2 = x; }
-   inline void SetY(int y) { mY = y; }
+   inline void SetX(float x) { mX = x; }
+   inline void SetX2(float x) { mX2 = x; }
+   inline void SetY(float y) { mY = y; }
 
    // Methods
    virtual void Update (float dt);
